+++ src/love/image.h
@@ -15,15 +15,6 @@ class image {
 	public:
 	bool load();
 	bool unload();
-
-	/**
-	 * Creates a new ImageData object.
-	 *
-	 * @param filename The filename of the image file.
-	 *
-	 * @return The new ImageData object.
-	 */
-	Image* newImageData(const std::string& filename);
 };
 
 }  // namespace love