+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - No Game demo when ChaiLove is loaded without content
 - Version mismatch warning when content's ChaiLove version doesn't match core's
 - `love.data.hash()` `sha1` support
+- Replaced `love.image.newImageData` with `love.graphics.newImage()`
 
 ## 0.19.0 - 2018-04-13
 ### Added